~ abstract test reporter

+++ test/flood-test.js
@@ -3,8 +3,8 @@
 
 const QueueSize = 'PROMISE_QUEUE_SIZE' in process.env
       ? parseInt(process.env.PROMISE_QUEUE_SIZE)
-      : 25
-const LogProcessStats = true
+      : 5
+const LogProcessStats = false // show summar at end
 
 const child_process = require('child_process')
 const AllTests = []
@@ -12,19 +12,33 @@ const Queue = require('../timeout-promise-queue.js').PromiseQueue(QueueSize)
 const _AfterAllTests = typeof jest !== 'undefined' ? afterAll : after
 const TestTimeoutMessage = 'timeout exceeded blah bah blah'
 
-/* start processes */
+// Test list:
 Tests = [
+  // Lead with long sleeps so Queue doesn't deplete under mocha.
   { processes: 10, sleep: 10000, timeout: 20000, ok: true },
+
+  { processes: 1, sleep: 1000, ok: true },
+
+  { processes: 1, sleep: 10000, timeout: 20, ok: false,
+    // no rejection parameter so expect timeout-promise-queue's default:
+    exceptionPattern: RegExp('^timeout of 20 exceeded$') },
+
+  { processes: 1, sleep: 10000, timeout: 20, ok: false,
+    rejection: Error(TestTimeoutMessage),
+    exceptionPattern: RegExp('^' + TestTimeoutMessage + '$') },
+
   { processes: 10, sleep: 10000, timeout: 200, ok: false,
     rejection: () => Error(TestTimeoutMessage),
     exceptionPattern: RegExp('^' + TestTimeoutMessage + '$') }
 ]
 let SuiteTimeout = 250 // generous setup and tear-down time
+let ExpectedQueueSize = -QueueSize // first QueueSize will dispatch immediately.
 Tests.forEach((test, idx) => {
   const command = 'setTimeout(() => { process.exit(0); }, ' + test.sleep + ')'
   // const Command = 'for (let i = 0; i < 2**28; ++i) ;' // The busy alternative.
   SuiteTimeout += 2 * test.sleep * test.processes / QueueSize
   startProcesses(idx, test.processes, command, test.timeout, test.ok, test.rejection)
+  ExpectedQueueSize += test.processes
 })
 
 function startProcesses (batch, processes, command, timeout, ok, rejection) {
@@ -38,49 +52,56 @@ function startProcesses (batch, processes, command, timeout, ok, rejection) {
         let program = child_process.spawn('node', ['-e', command])
         program.on('exit', exitCode => { resolve({exitCode:exitCode}) })
         program.on('error', reject)
-        cancel.on('timeout', err => {
-          program.kill()
-          reject()
-        })
+        if (cancel)
+          cancel.on('timeout', err => {
+            program.kill()
+            reject()
+          })
       }), timeout, rejection)
     })
   }
 }
 
 /* test results */
-describe('churn', () => {
+describe('timeout-promise-queue', () => {
+  it('should have the right size', () => {
+    if (Queue.size() !== ExpectedQueueSize)
+      throw Error('expected ' + Queue.size() +' === ' + ExpectedQueueSize)
+  })
   AllTests.forEach(test => {
     let title = 'should ' + (test.ok ? 'pass' : 'fail') + ' test ' + test.label + '.'
     function cb (done) {
       test.exec.then(exec => {
-        test.end = new Date()
         if (test.ok) {
-          test.message = 'OK'
-          done()
+          report(true)
         } else {
-          test.message = 'unexpected success ' + JSON.stringify(test)
-          done(Error(test.message))
+          report(false, 'unexpected success')
         }
       }).catch(e => {
-        test.end = new Date()
         if (test.ok) {
-          test.message = e
-          done(e)
+          report(false, e.message)
         } else {
           if (test.exceptionPattern) {
             if (e.message.match(test.exceptionPattern)) {
-              test.message = 'OK'
-              done()
+              report(true)
             } else {
-              test.message = "expected \"" + e.message + "\" to match /" + test.exceptionPattern + "/"
-              done(Error(test.message))
+              report(false, "expected \"" + e.message + "\" to match exceptionPattern")
             }
           } else {
-            test.message = 'OK'
-            done()
+            report(true)
           }
         }
       })
+
+      function report (pass, message) {
+        test.end = new Date()
+        test.message = message || (pass ? 'OK' : 'FAIL')
+        if (pass) {
+          done()
+        } else {
+          done(Error(message + ' ' + JSON.stringify(test)))
+        }
+      }
     }
     if (typeof jest !== 'undefined') {
       it(title, cb, SuiteTimeout)
